+++ index.html
@@ -235,7 +235,10 @@
     </section>
   </template>
 
-  <script src="./js/main.js" defer></script>
+  <script src="./js/main.js" type="module" defer></script>
+  <script src="./js/util.js" type="module" defer></script>
+  <script src="./js/photo-comments.js" type="module" defer></script>
+  <script src="./js/photo-discriptions.js" type="module" defer></script>
   <script src="./js/functions.js" defer></script>
 
 </body>
